# frontend/packages/core/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Re:Pibot Shop API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/me/gifts/redeem": {
      "post": {
        "tags": [
          "subscription"
        ],
        "summary": "Redeem Gift",
        "operationId": "redeem_gift_api_me_gifts_redeem_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemGiftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/orders": {
      "get": {
        "tags": [
          "subscription"
        ],
        "summary": "List Orders",
        "operationId": "list_orders_api_me_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderResponse"
                  },
                  "type": "array",
                  "title": "Response List Orders Api Me Orders Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "subscription"
        ],
        "summary": "Create Order",
        "operationId": "create_order_api_me_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/orders/{order_id}": {
      "get": {
        "tags": [
          "subscription"
        ],
        "summary": "Get Order",
        "operationId": "get_order_api_me_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/subscription/auto-renew": {
      "get": {
        "tags": [
          "subscription"
        ],
        "summary": "Get Auto Renew",
        "operationId": "get_auto_renew_api_me_subscription_auto_renew_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutoRenewResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "subscription"
        ],
        "summary": "Set Auto Renew",
        "operationId": "set_auto_renew_api_me_subscription_auto_renew_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoRenewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutoRenewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/devices/unlink": {
      "post": {
        "tags": [
          "subscription"
        ],
        "summary": "Unlink Device",
        "operationId": "unlink_device_api_me_devices_unlink_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnlinkDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutoRenewRequest": {
        "properties": {
          "auto_renew_enabled": {
            "type": "boolean",
            "title": "Auto Renew Enabled"
          }
        },
        "type": "object",
        "required": [
          "auto_renew_enabled"
        ],
        "title": "AutoRenewRequest"
      },
      "AutoRenewResponse": {
        "properties": {
          "auto_renew_enabled": {
            "type": "boolean",
            "title": "Auto Renew Enabled"
          }
        },
        "type": "object",
        "required": [
          "auto_renew_enabled"
        ],
        "title": "AutoRenewResponse"
      },
...
      "CreateOrderRequest": {
        "properties": {
          "plan_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Plan Id"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "purchase",
              "renew",
              "gift"
            ],
            "title": "Purpose"
          },
          "provider": {
            "type": "string",
            "enum": [
              "yookassa",
              "stars"
            ],
            "title": "Provider"
          },
          "promo_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Promo Code"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Idempotency Key"
          },
          "save_payment_method": {
            "type": "boolean",
            "title": "Save Payment Method",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "purpose",
          "provider",
          "idempotency_key"
        ],
        "title": "CreateOrderRequest",
        "description": "Только намерение покупателя; тариф и сумма всегда читаются сервером."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "plan_id": {
            "type": "integer",
            "title": "Plan Id"
          },
          "plan_code": {
            "type": "string",
            "title": "Plan Code"
          },
          "plan_name": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Plan Name"
          },
          "duration_days": {
            "type": "integer",
            "title": "Duration Days"
          },
          "price_rub": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price Rub"
          },
          "price_stars": {
            "type": "integer",
            "title": "Price Stars"
          },
          "gross_rub": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Gross Rub"
          },
          "discount_rub": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Discount Rub"
          },
          "amount_due_rub": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount Due Rub"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "confirmation_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Url"
          },
          "telegram_invoice_required": {
            "type": "boolean",
            "title": "Telegram Invoice Required",
            "default": false
          },
          "telegram_handoff_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Handoff Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "purpose",
          "plan_id",
          "plan_code",
          "plan_name",
          "duration_days",
          "price_rub",
          "price_stars",
          "gross_rub",
          "discount_rub",
          "amount_due_rub",
          "status",
          "expires_at",
          "confirmation_url"
        ],
        "title": "OrderResponse",
        "description": "Без provider payload: клиенту достаточно снимка заказа и URL оплаты."
      },
...
      "RedeemGiftRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "RedeemGiftRequest"
      },
...
      "SubscriptionResponse": {
        "properties": {
          "plan_code": {
            "type": "string",
            "title": "Plan Code"
          },
          "plan_name": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Plan Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "subscription_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Url"
          },
          "traffic_limit_bytes": {
            "type": "integer",
            "title": "Traffic Limit Bytes"
          },
          "hwid_device_limit": {
            "type": "integer",
            "title": "Hwid Device Limit"
          },
          "auto_renew_enabled": {
            "type": "boolean",
            "title": "Auto Renew Enabled"
          }
        },
        "type": "object",
        "required": [
          "plan_code",
          "plan_name",
          "status",
          "started_at",
          "expires_at",
          "subscription_url",
          "traffic_limit_bytes",
          "hwid_device_limit",
          "auto_renew_enabled"
        ],
        "title": "SubscriptionResponse"
      },
      "SubscriptionStateResponse": {
        "properties": {
          "subscription": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubscriptionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "trial_available": {
            "type": "boolean",
            "title": "Trial Available"
          }
        },
        "type": "object",
        "required": [
          "subscription",
          "trial_available"
        ],
        "title": "SubscriptionStateResponse",
        "description": "Подписка вместе с правом на триал.\n\nОдним ответом, а не двумя запросами: экран подписки решает по обоим полям\nсразу, показать срок или кнопку триала."
      },
...
      "UnlinkDeviceRequest": {
        "properties": {
          "hwid": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Hwid"
          }
        },
        "type": "object",
        "required": [
          "hwid"
        ],
        "title": "UnlinkDeviceRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}